Move parse.go to internal directory
