+++ jsv2/jetstream/message.go
@@ -21,7 +21,7 @@ import (
 	"time"
 
 	"github.com/nats-io/nats.go"
-	"github.com/nats-io/nats.go/parser"
+	"github.com/nats-io/nats.go/internal/parser"
 )
 
 type (
